+++ .env.example
@@ -33,6 +33,9 @@ UVICORN_PORT = 8000
 # CLASH_SUBSCRIPTION_TEMPLATE="clash/my-custom-template.yml"
 # SUBSCRIPTION_PAGE_TEMPLATE="subscription/index.html"
 # HOME_PAGE_TEMPLATE="home/index.html"
+# V2RAY_SUBSCRIPTION_TEMPLATE="v2ray/default.json"
+# SINGBOX_SUBSCRIPTION_TEMPLATE="singbox/default.json"
+# MUX_TEMPLATE="mux/default.json"
 
 # Set headers for subscription
 # SUB_PROFILE_TITLE = "Susbcription"
