+++ app/db/crud.py
@@ -82,7 +82,11 @@ def update_hosts(db: Session, inbound_tag: str, modified_hosts: List[ProxyHostMo
             alpn=host.alpn,
             fingerprint=host.fingerprint,
             allowinsecure=host.allowinsecure,
-            is_disabled=host.is_disabled
+            is_disabled=host.is_disabled,
+            sockopt=host.sockopt,
+            sockopt_enable=host.sockopt_enable,
+            proxy_outbound=host.proxy_outbound,
+            fragment_setting=host.fragment_setting,
         ) for host in modified_hosts
     ]
     db.commit()
