+++ app/models/proxy.py
@@ -2,6 +2,7 @@ import json
 from enum import Enum
 from typing import Optional, Union
 from uuid import UUID, uuid4
+import re
 
 from pydantic import BaseModel, Field, validator
 
@@ -15,6 +16,7 @@ from xray_api.types.account import (
     XTLSFlows,
 )
 
+FRAGMENT_PATTERN = re.compile(r'^(\d{1,3}-\d{1,3}),(\d{1,3}-\d{1,3}),tlshello')
 
 class ProxyTypes(str, Enum):
     # proxy_type = protocol
@@ -141,6 +143,11 @@ class ProxyHost(BaseModel):
     fingerprint: ProxyHostFingerprint = ProxyHostFingerprint.none
     allowinsecure: Union[bool, None] = None
     is_disabled: Union[bool, None] = None
+    sockopt: Union[dict, None] = None
+    sockopt_enable: Union[bool, None] = None
+    proxy_outbound: Union[dict, None] = None
+    mux_enable: Union[bool, None] = None
+    fragment_setting: Optional[str] = Field(None, nullable=True)
 
     class Config:
         orm_mode = True
@@ -161,6 +168,14 @@ class ProxyHost(BaseModel):
         except ValueError as exc:
             raise ValueError("Invalid formatting variables")
 
+        return v
+    
+    @validator("fragment_setting", check_fields=False)
+    def validate_fragment(cls, v):
+        if v and not FRAGMENT_PATTERN.match(v):
+            raise ValueError(
+                "Fragment setting must be like this: length,interval,packet (10-100,100-200,tlshello)."
+            )
         return v
