+++ app/templates/mux/default.json
@@ -1,0 +1,13 @@
+{
+    "v2ray": {
+          "enabled": true,
+          "concurrency": 8,
+          "xudpConcurrency": 8,
+          "xudpProxyUDP443": "reject"
+        },
+    "sing-box": {
+          "enabled": false,
+          "protocol": "smux",
+          "max_streams": 8
+        }
+  }
