--- app/templates/singbox/mux_config.json
@@ -1,5 +1,0 @@
-{
-    "enabled": false,
-    "protocol": "smux",
-    "max_streams": 8
-}
